Network Security Testing/requirements.txt
openai==0.28.0
reportlab
markdown2
beautifulsoup4
weasyprint
